Seperate out test and comment out older test
